+++ src/Compilers/CSharp/Portable/FlowAnalysis/PreciseAbstractFlowPass_Switch.cs
@@ -1,19 +1,10 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using Microsoft.CodeAnalysis.Collections;
 using Microsoft.CodeAnalysis.CSharp.Symbols;
-using Microsoft.CodeAnalysis.Text;
-using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CSharp
 {
-    internal abstract partial class PreciseAbstractFlowPass<TLocalState>
+    internal abstract partial class AbstractFlowPass<TLocalState>
     {
         #region implementation for the old-style (no-patterns) variation of the switch statement.
 
@@ -166,7 +157,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                             SetState(StateWhenTrue);
                         }
 
-                        _pendingBranches.Add(new PendingBranch(label, this.State));
+                        PendingBranches.Add(new PendingBranch(label, this.State));
                     }
                 }
             }
@@ -184,7 +175,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                     SetState(initialState.Clone());
                 }
 
-                _pendingBranches.Add(new PendingBranch(node.DefaultLabel, this.State));
+                PendingBranches.Add(new PendingBranch(node.DefaultLabel, this.State));
             }
 
             // visit switch sections
